JP - Converting code to use Python3

+++ resources/lib/xbmcvfs.py
@@ -26,8 +26,8 @@ class File(object):
         f = xbmcvfs.File(file, 'w')
     """
     fh = None
-    
-    
+
+
     def __init__(self, filepath, mode=None):
         """
         :param filepath: file path
@@ -101,7 +101,7 @@ class File(object):
             result = f.seek(8129, 0)
             f.close()
         """
-        return long()
+        return int()
 
     def size(self):
         """
@@ -113,7 +113,7 @@ class File(object):
             s = f.size()
             f.close()
         """
-        return long()
+        return int()
 
     def write(self, buffer):
         """
@@ -230,13 +230,13 @@ def listdir(path):
         dirs, files = xbmcvfs.listdir(path)
     """
     #path = os.path.normpath(path)
-        
+
     dirs = [d for d in os.listdir(path) if os.path.isdir(os.path.join(path, d))]
     #print 'dirs: %s' %dirs
     files = [f for f in os.listdir(path) if os.path.isfile(os.path.join(path, f))]
     #print 'files: %s' %files
-    
-    return dirs, files 
+
+    return dirs, files
 
 
 class Stat(object):
@@ -263,28 +263,28 @@ class Stat(object):
         """
 
     def st_mode(self):
-        return long()
+        return int()
 
     def st_ino(self):
-        return long()
+        return int()
 
     def st_nlink(self):
-        return long()
+        return int()
 
     def st_uid(self):
-        return long()
+        return int()
 
     def st_gid(self):
-        return long()
+        return int()
 
     def st_size(self):
-        return long()
+        return int()
 
     def st_atime(self):
-        return long()
+        return int()
 
     def st_mtime(self):
-        return long()
+        return int()
 
     def st_ctime(self):
-        return long()
+        return int()
